+++ lib/src/widgets/toolbar/buttons/select_header_style_button.dart
@@ -20,6 +20,7 @@ class QuillToolbarSelectHeaderStyleButton extends StatefulWidget {
   });
 
   final QuillController controller;
+  // TODO: Needs to be reviewed
   final QuillToolbarSelectHeaderStyleButtonsOptions options;
 
   @override